Added 2GL base class for D1-2GL, D2, PD2 to reduce code duplication.
