+++ packages/zoltan2/core/src/algorithms/color/Zoltan2_ColoringAlgorithms.hpp
@@ -48,6 +48,7 @@
 #include <Zoltan2_Standards.hpp>
 #include <Zoltan2_AlgSerialGreedy.hpp>
 #include <Zoltan2_AlgHybridD1.hpp>
+#include <Zoltan2_AlgHybrid2GL.hpp>
 #include <Zoltan2_AlgHybridD1-2GL.hpp>
 #include <Zoltan2_AlgHybridD2.hpp>
 #include <Zoltan2_AlgHybridPD2.hpp>
